Add null checking to ServletLibrary functions

+++ project/src/main/java/com/google/sps/data/ServletLibrary.java
@@ -30,6 +30,10 @@ public class ServletLibrary {
   // Returns a product set object with all the information stored in the database.
   // If there is no product set, returns null. 
   public static ProductSetEntity retrieveProductSetInfo(DatastoreService datastore, String inputQuery, boolean inputIsDisplayName) {
+    if (datastore == null) {
+      System.err.println("Datastore is null!");
+      return null;
+    }
     Filter filter = new FilterPredicate("productSetId", FilterOperator.EQUAL, inputQuery);
     if (inputIsDisplayName) {
       filter = new FilterPredicate("productSetDisplayName", FilterOperator.EQUAL, inputQuery);
@@ -62,6 +66,14 @@ public class ServletLibrary {
   // Returns an Account object with all the information that is stored in datstore.
   // If the account is not in datastore, returns null.
   public static Account retrieveAccountInfo(DatastoreService datastore, UserService userService, String userId) {
+    if (datastore == null) {
+      System.err.println("Datastore is null!");
+      return null;
+    }
+    if (userService == null) {
+      System.err.println("UserService is null!");
+      return null;
+    }
     Filter filter = new FilterPredicate("userId", FilterOperator.EQUAL, userId);
     Query query = new Query("Account").setFilter(filter);
     PreparedQuery pq = datastore.prepare(query);
@@ -123,6 +135,10 @@ public class ServletLibrary {
 
   // Returns a list of product set objects, taken from datastore.
   public static List<ProductSetEntity> listAllProductSets(DatastoreService datastore) {
+    if (datastore == null) {
+      System.err.println("Datastore is null!");
+      return null;
+    }
     Query query = new Query("ProductSet").addSort("productSetDisplayName", SortDirection.ASCENDING);
     PreparedQuery pq = datastore.prepare(query);
     List<ProductSetEntity> results = new ArrayList<>();
@@ -152,6 +168,11 @@ public class ServletLibrary {
   // Returns a business object containing all the business information stored in 
   // datastore.
   public static Business retrieveBusinessInfo(DatastoreService datastore, String businessId) {
+    if (datastore == null) {
+      System.err.println("Datastore is null!");
+      return null;
+    }
+
     // Retrieving from datastore.
     Filter filter = new FilterPredicate("businessId", FilterOperator.EQUAL, businessId);
     Query query = new Query("Business").setFilter(filter);
@@ -212,6 +233,10 @@ public class ServletLibrary {
   // For every label that a product has, we assign the product to that label in
   // the labels table in datastore.
   public static void addProductToLabels(DatastoreService datastore, String productId, List<String> labels) {
+    if (datastore == null || productId == null || labels == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }
     for (String label : labels) {
       Filter filter = new FilterPredicate("label", FilterOperator.EQUAL, label.toLowerCase());
       Query query = new Query("ProductLabel").setFilter(filter);
@@ -240,6 +265,10 @@ public class ServletLibrary {
   public static void deleteProductFromLabels(DatastoreService datastore, 
                                              String productId, 
                                              List<String> labels) {
+    if (datastore == null || productId == null || labels == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }                                          
     for (String label : labels) {
       Filter filter = new FilterPredicate("label", FilterOperator.EQUAL, label.toLowerCase());
       Query query = new Query("ProductLabel").setFilter(filter);
@@ -262,6 +291,10 @@ public class ServletLibrary {
                                          String productId,
                                          List<String> oldLabels,
                                          List<String> labels) {
+    if (datastore == null || productId == null || labels == null || oldLabels == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }
     // Check what needs to be deleted and added.
     Set<String> oldLabelsSet = new HashSet<>(oldLabels);
     Set<String> labelsSet = new HashSet<>(labels);
@@ -281,6 +314,10 @@ public class ServletLibrary {
 
   // Add product to the specified product set.
   public static void addProductToProductSet(DatastoreService datastore, String productId, String productSetId) {
+    if (datastore == null || productId == null || productSetId == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }
     Filter filter = new FilterPredicate("productSetId", FilterOperator.EQUAL, productSetId);
     Query query = new Query("ProductSet").setFilter(filter);
     PreparedQuery pq = datastore.prepare(query);
@@ -299,6 +336,10 @@ public class ServletLibrary {
 
   // Delete product from the specifiec set.
   public static void deleteProductFromProductSet(DatastoreService datastore, String productId, String productSetId) {
+    if (datastore == null || productId == null || productSetId == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }
     Filter filter = new FilterPredicate("productSetId", FilterOperator.EQUAL, productSetId);
     Query query = new Query("ProductSet").setFilter(filter);
     PreparedQuery pq = datastore.prepare(query);
@@ -317,6 +358,10 @@ public class ServletLibrary {
                                        String productId, 
                                        String oldProductSetId,
                                        String productSetId) {
+    if (datastore == null || productId == null || productSetId == null || oldProductSetId == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }
     if (oldProductSetId.equals(productSetId)) return;
     deleteProductFromProductSet(datastore, productId, oldProductSetId);
     addProductToProductSet(datastore, productId, productSetId);
@@ -324,6 +369,10 @@ public class ServletLibrary {
 
   // Add product to the specified product category.
   public static void addProductToProductCategory(DatastoreService datastore, String productId, String productCategory) {
+    if (datastore == null || productId == null || productCategory == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }
     Filter filter = new FilterPredicate("productCategory", FilterOperator.EQUAL, productCategory);
     Query query = new Query("ProductCategory").setFilter(filter);
     PreparedQuery pq = datastore.prepare(query);
@@ -351,6 +400,10 @@ public class ServletLibrary {
   public static void deleteProductFromProductCategory(DatastoreService datastore, 
                                                       String productId, 
                                                       String productCategory) {
+    if (datastore == null || productId == null || productCategory == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }                                                  
     Filter filter = new FilterPredicate("productCategory", FilterOperator.EQUAL, productCategory);
     Query query = new Query("ProductCategory").setFilter(filter);
     PreparedQuery pq = datastore.prepare(query);
@@ -370,6 +423,11 @@ public class ServletLibrary {
                                        String productId, 
                                        String oldProductCategory,
                                        String productCategory) {
+    if (datastore == null || productId == null || productCategory == null || 
+        oldProductCategory == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }
     if (oldProductCategory.equals(productCategory)) return;
     deleteProductFromProductCategory(datastore, productId, oldProductCategory);
     addProductToProductCategory(datastore, productId, productCategory);
@@ -377,6 +435,10 @@ public class ServletLibrary {
 
   // Add product to the list of products offerec by the business.
   public static void addProductToBusiness(DatastoreService datastore, String productId, String businessId) {
+    if (datastore == null || productId == null || businessId == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }
     Filter filter = new FilterPredicate("businessId", FilterOperator.EQUAL, businessId);
     Query query = new Query("Business").setFilter(filter);
     PreparedQuery pq = datastore.prepare(query);
@@ -395,12 +457,16 @@ public class ServletLibrary {
 
   // Delete a given product from the business.
   public static void deleteProductFromBusiness(DatastoreService datastore, String productId, String businessId) {
+    if (datastore == null || productId == null || businessId == null) {
+      System.err.println("At least one of the inputs was null!");
+      return;
+    }
     Filter filter = new FilterPredicate("businessId", FilterOperator.EQUAL, businessId);
     Query query = new Query("Business").setFilter(filter);
     PreparedQuery pq = datastore.prepare(query);
     Entity entity = pq.asSingleEntity();
     if (entity != null) {
-      @SuppressWarnings("unchecked") // Documentation says suppress warning this way
+      @SuppressWarnings("unchecked") // Documentation says to suppress warning this way
         List<String> productIds = (ArrayList<String>) entity.getProperty("productIds"); 
       if (productIds == null) productIds = new ArrayList<String>();
       productIds.remove(productId);
@@ -417,6 +483,10 @@ public class ServletLibrary {
                                                  String productCategory,
                                                  String sortOrder,
                                                  String textQuery) {
+    if (datastore == null) {
+      System.err.println("Datastore was null!");
+      return null;
+    }
     // Set the filters.
     List<Filter> filters = new ArrayList<>();
     if (businessId != null) {
@@ -541,6 +611,10 @@ public class ServletLibrary {
 
   // Retrieves product information based on the product id. 
   public static ProductEntity retrieveProductInfo(DatastoreService datastore, String productId) {
+    if (datastore == null || productId == null) {
+      System.err.println("At least one of the inputs was null!");
+      return null;
+    }
     // Retrieving from datastore.
     Filter filter = new FilterPredicate("productId", FilterOperator.EQUAL, productId);
     Query query = new Query("Product").setFilter(filter);
